Add poor man's "show most recent tag" by using git-describe
In the diff view, the "Refs:" lines will now sport the output of
git-describe by appending it as the last reference to the line.

Now, it should probably be optional and I mostly added it because I wanted
to do a little tig hacking. Anyway, let's hope it will be useful.