Supply explicit permission bits to 'install'
'install' will install files with permissions set to '0755' if the
permissions are not specified on the command line.  Having the execute
bits set on non-executable content is not desireable.  Specify mode
0644 for non-executable content as well as specify mode 0755 for
executable content (in case the defaults change or are different on
different systems).  Also tell 'install' to preserve timestamps.

[ Updated to also do it when installing HTML files. --jonas ]

Signed-off-by: Jeffrey C. Ollie <jeff@ocjtech.us>
Signed-off-by: Jonas Fonseca <fonseca@diku.dk>