Make it possible to overwrite the default (terminal) colors
Use the setting of the "default" color pair to set the default colors of
the terminal via assume_default_colors(). Fall back to white on black.
With this it is possible to no longer use the terminal colors if
desired.

This was requested by Sebastian Harl as a way to fix Debian bug 399884.