[GH #32] Only color +/- in the second column when in a combined diff
If the diff type is not combined, force the mistakenly identified lines
to the default line type.

Combined diffs are used by git when showing changes on a merge commit.
Refer to the section COMBINED DIFF FORMAT in man git-diff.

This resolves [GH #32] for the diff and stage views.  The general pager
view still exhibits the issue when used like so `git show -p | tig`, but
this seems to be a less common use case that needs additional state to
fix.