Parse commit title correctly in presence of mergetag object
When merging a signed tag, recent versions of Git store the original tag
in the merge commit as a "mergetag object". This additional content can
confuse tig's parsing of the commit title for the main view if the tag
description contains a line which happens to start with four spaces.
(Recent examples of such commits in linux.git: 94fb175c04, 9479f0f801,
1ddca05743.)

Git helpfully separates with a blank line the commit header (which
contains the mergetag object) from the description itself, so keep track
of which lines belong to the header and only look for the title in the
description.

Signed-off-by: Romain Francoise <romain@orebokech.com>
Signed-off-by: Jonas Fonseca <fonseca@diku.dk>