Implement navigating back to previous view state in the blame view
Use the new view history API to track the previous blamed revisions when
navigating to the blame information for the current file. Data for the
current blame view is stored inside the view state and pushed onto the
stack when reloading another blamed revision from inside the blame view.

This also improves error messages when trying to reloading the blamed
revision currently displayed as well as navigating back in views that
does not support it.

Note, that the back action (REQ_BACK) acts similar to REQ_PARENT in the
tree view.

Fixes #124
Fixes #139