Fix vertical split regression by hardcoding the split parameter
This fixes vertical splitting by turning opt_scale_vsplit_view into the
VSPLIT_SCALE #define which hard-codes the width scale to 0.5. The
regression was introduced when the default option values where moved to
tigrc and the value for opt_scale_vsplit_view was removed. This caused
the vertical split to only asign a width of 4 columns to the right-most
view.