remove `--boundary` from `%(cmdlineargs)`
If you ask tig to show boundary commits, like:

  tig --boundary origin...master

we end up passing `--boundary` along to `git show` when we
open the diff view for the commit. As a result, each diff
contains not only the commit we cared about, but also the
commit message and diff for its parent.

We can fix this by adding `--boundary` to the list of
revision args to omit in `filter_options`.