Improve auto-refreshing of views to use various update info
This adds hooks in different parts of the update code which acts as
sources for watching for state changes in the views. As an example
index changes are now tracked in more detail and diffed to improve
updating of the stage, main, and status views.

It also registers watches for HEAD and references in the main and refs
view so that both views are refreshed after for example a branch is
deleted.