Use buffer for reading view data
Change the view update code to pass the buffer from io_get() to the view
read methods and use the buffer when extracting view data via
io_memchr(). This fixes views that use NUL-terminated input to work
correctly when data is encoding, for example when "gui.encoding" is set.

Fixes #287