Partly restore old refresh behavior
This ensures that the current view is refreshed after returning from an
external command when `refresh-mode` is set to `after-command`. That was
what Tig used to do before the option was introduced. It special cases
the 'after-command' event to force a refresh in the current view.

The fix is tested to work in both the main and status view.

Fixes #289