Fix and improve status on-branch information
Add a test for the different on-branch "modes" based on whether
git-rebase, git-am or git-merge is in progress or whether a detached
head has been checked out.

This also changes status view refreshing to only reload the current
branch information and not all refs as suggested in #341. Note that it
actually reloads all repo information to get both the symbolic ref
pointed to by HEAD as well as the HEAD id.