ui: fix crash due to out-of-bounds array access
If the search string typed at the "Find file:" prompt in the file finder
does not match any files, then the resulting file list is empty. When
ENTER is pressed, open_file_finder() attempts to retrieve the
non-existent entry at index 0 in the empty list, often resulting in a
crash. Fix this boundary condition error.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>
Signed-off-by: Jonas Fonseca <jonas.fonseca@gmail.com>