make tig the process-group leader
and give it the foreground connection to the TTY.  This announces to
the OS "I am an interactive program taking input".  In the most common
case, this will already have been arranged by the user's shell.

This init is logically done as early as possible, as child processes
inherit these characteristics.

This also allows tig to easily ensure child process cleanup at exit.