Restore support for tig log --graph
Support for `tig log --graph` was added in 2.0 with fa2fcc9 but was
later broken in 2.2 with 178c45c when the box structure was
implemented. This makes it work again, though it would be cleaner
to move the implementation from draw.c to log.c and pager.c,
using the box structure.