Makefile: avoid overzealous removal of compile_commands.json (#1157)
Our logic to generate a compilation database is copied (almost
verbatim) from Git's Makefile,

However, one difference is that before creating compile_commands.json
we always remove it.
This doesn't seem very useful, plus it thwarts the logic two lines
below to overwrite compile_commands.json only if we successfully
created a non-empty file.  I believe I added this while debugging
this logic.

I'm not sure in which scenarios this actually makes a difference,
but it seems wrong to keep it.